ramsay/requirements.txt
django >= 2.2.13, < 3.0
djangorestframework
markdown
django-filter
pillow
